src/scrapers/requirements.txt
# Core dependencies
scrapy==2.11.1
scrapy-playwright==0.0.34
playwright==1.50.0
pymongo==4.6.1
python-dotenv==1.0.0

# Proxy and User Agent
scrapy-rotating-proxies==0.6.2
fake-useragent==1.4.0

# Data Processing
itemadapter==0.11.0
w3lib==2.3.1
python-dateutil==2.8.2
tldextract==5.1.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0

# Monitoring
prometheus-client==0.20.0
statsd==4.0.1

# Development
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 